Add host-side test harness; fix settings bounds, override wraparound, and arc UB
test/ (Phase 4.1 — host-side test harness):
- Compiles the real gcode.c, planner.c, and nuts_bolts.c natively (host
  gcc, no AVR toolchain needed) against a minimal AVR shim
  (test/avr_shim/) and a stub layer (test/grbl_stubs.c) standing in for
  the hardware-facing modules these three call into.
- 39 test cases across the three files. Run via `make test` (and
  `make test-clean`), independent of the AVR build targets.
- See test/README.md for exact scope, what's stubbed vs. really
  reimplemented, and a note on the harness having been verified to
  actually catch a deliberately-injected regression.
- Updated CLAUDE.md's Testing section, which previously (and now
  incorrectly) stated there was no host-side harness.

grbl/settings.c, grbl/report.h, doc/csv/error_codes_en_US.csv:
- settings_store_global_setting() now rejects values <= 0 for
  steps/mm, max rate, acceleration, and arc tolerance. Previously only
  negative values were rejected, so e.g. $100=0 was accepted and later
  divided by zero in the planner and in every mm<->step conversion.
  Added STATUS_SETTING_VALUE_OUT_OF_RANGE (39) and its CSV row.

grbl/protocol.c:
- Feed and spindle override accumulation now uses int16_t instead of
  uint8_t while applying +/- increments, before the final min()/max()
  clamp. Previously, in any configuration where MIN_*_OVERRIDE exceeds
  the coarse increment, a single "decrease" request could underflow the
  uint8_t accumulator and have min() see the wrapped value as larger
  than MAX_*_OVERRIDE, snapping the override to maximum instead of the
  intended minimum -- verified with a standalone extraction of the
  exact arithmetic before applying the fix.

grbl/motion_control.c:
- mc_arc() now guards its segment-count formula against
  2*radius <= arc_tolerance. The unguarded formula took sqrt() of a
  value <= 0 for a sufficiently small arc, and converting the resulting
  NaN/Inf to uint16_t is undefined behavior in C -- confirmed by
  extracting the formula standalone, which produced 0 via UB on this
  host compiler with no guarantee any other compiler would agree.
  Confirmed zero change in segment count for four realistic arc sizes.

Verified throughout with the AVR toolchain bundled with the Arduino IDE
(avr-gcc 7.3.0 / avrdude 6.3.0): builds clean (only the pre-existing,
out-of-scope eeprom.c checksum warning), 29768->29930 bytes text
(+162 over the ~2494-byte Uno headroom), and `make test` passes 39/39.

+++ grbl/protocol.c
@@ -407,7 +407,13 @@ void protocol_exec_rt_system()
   if (rt_exec) {
     system_clear_exec_motion_overrides(); // Clear all motion override flags.
 
-    uint8_t new_f_override =  sys.f_override;
+    // NOTE: Accumulated in a signed 16-bit type, not uint8_t, so a decrement below
+    // MIN_FEED_RATE_OVERRIDE goes negative instead of wrapping around through 255.
+    // With stock defaults MIN_FEED_RATE_OVERRIDE (10) equals FEED_OVERRIDE_COARSE_INCREMENT
+    // (10), so the wrapped case never arose in practice -- but any config where MIN is
+    // greater than the increment would have had a single "decrease" request clamp to
+    // MAX_FEED_RATE_OVERRIDE instead of MIN, since min() saw a huge wrapped uint8_t value.
+    int16_t new_f_override = sys.f_override;
     if (rt_exec & EXEC_FEED_OVR_RESET) { new_f_override = DEFAULT_FEED_OVERRIDE; }
     if (rt_exec & EXEC_FEED_OVR_COARSE_PLUS) { new_f_override += FEED_OVERRIDE_COARSE_INCREMENT; }
     if (rt_exec & EXEC_FEED_OVR_COARSE_MINUS) { new_f_override -= FEED_OVERRIDE_COARSE_INCREMENT; }
@@ -435,7 +441,9 @@ void protocol_exec_rt_system()
     system_clear_exec_accessory_overrides(); // Clear all accessory override flags.
 
     // NOTE: Unlike motion overrides, spindle overrides do not require a planner reinitialization.
-    uint8_t last_s_override =  sys.spindle_speed_ovr;
+    // Accumulated in int16_t for the same reason as new_f_override above: a decrement
+    // below MIN_SPINDLE_SPEED_OVERRIDE must go negative, not wrap around through 255.
+    int16_t last_s_override = sys.spindle_speed_ovr;
     if (rt_exec & EXEC_SPINDLE_OVR_RESET) { last_s_override = DEFAULT_SPINDLE_SPEED_OVERRIDE; }
     if (rt_exec & EXEC_SPINDLE_OVR_COARSE_PLUS) { last_s_override += SPINDLE_OVERRIDE_COARSE_INCREMENT; }
     if (rt_exec & EXEC_SPINDLE_OVR_COARSE_MINUS) { last_s_override -= SPINDLE_OVERRIDE_COARSE_INCREMENT; }
